requirements.txt
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-SQLAlchemy==2.4.3
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
Pillow==7.1.2
requests==2.24.0
SQLAlchemy==1.3.17
urllib3==1.25.9
uuid==1.30
Werkzeug==1.0.1
